backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
passlib==1.7.4
bcrypt==3.2.2
email-validator
python-jose[cryptography]
python-dotenv
pydantic-settings
python-multipart
pytest
httpx